+++ text_frame.go
@@ -41,7 +41,7 @@ func (tf TextFrame) Size() int {
 	return 1 + len(tf.Text)
 }
 
-func (tf TextFrame) WriteTo(w io.Writer) (n int, err error) {
+func (tf TextFrame) WriteTo(w io.Writer) (n int64, err error) {
 	var i int
 	bw := bwpool.Get(w)
 	defer bwpool.Put(bw)
@@ -56,7 +56,7 @@ func (tf TextFrame) WriteTo(w io.Writer) (n int, err error) {
 	if err != nil {
 		return
 	}
-	n += i
+	n += int64(i)
 
 	err = bw.Flush()
 	return
